plugins moved from tests dir to main src dir.
